added card service
- moved internal class in api controller to external file
