+++ www/client/plugins/filtreapi/HTTPReponseHelper.class.php
@@ -1,0 +1,173 @@
+<?php 
+
+class HTTPReponseHelper
+{
+
+    static $OK = 200;
+
+    static $BAD_REQUEST = 400;
+
+    static $UNAUTHORIZED = 401;
+
+    static $PAYMENT_REQUIRED = 402;
+
+    static $METHOD_NOT_ALLOWED = 402;
+
+    static $FORBIDDEN = 403;
+
+    static function getErrorJson($code, $message = null, $stringify = false)
+    {
+        $e = new stdClass();
+        if (! $message) {
+            $message = self::getMessage($code);
+        }
+        $e->code = $code;
+        $e->message = $message;
+        if ($stringify)
+            return json_encode($e);
+        return $e;
+    }
+
+    static function getResultError($code, $message = null)
+    {
+        return new Response(false, self::getErrorJson($code, $message));
+    }
+
+    static function getResult($success, $body = null)
+    {
+        return new Response($success, $body);
+    }
+
+    static function getMessage($code)
+    {
+        $text = 'Unknown http status code "' . $code . '"';
+        switch ($code) {
+            case 100:
+                $text = 'Continue';
+                break;
+            case 101:
+                $text = 'Switching Protocols';
+                break;
+            case 200:
+                $text = 'OK';
+                break;
+            case 201:
+                $text = 'Created';
+                break;
+            case 202:
+                $text = 'Accepted';
+                break;
+            case 203:
+                $text = 'Non-Authoritative Information';
+                break;
+            case 204:
+                $text = 'No Content';
+                break;
+            case 205:
+                $text = 'Reset Content';
+                break;
+            case 206:
+                $text = 'Partial Content';
+                break;
+            case 300:
+                $text = 'Multiple Choices';
+                break;
+            case 301:
+                $text = 'Moved Permanently';
+                break;
+            case 302:
+                $text = 'Moved Temporarily';
+                break;
+            case 303:
+                $text = 'See Other';
+                break;
+            case 304:
+                $text = 'Not Modified';
+                break;
+            case 305:
+                $text = 'Use Proxy';
+                break;
+            case 400:
+                $text = 'Bad Request';
+                break;
+            case 401:
+                $text = 'Unauthorized';
+                break;
+            case 402:
+                $text = 'Payment Required';
+                break;
+            case 403:
+                $text = 'Forbidden';
+                break;
+            case 404:
+                $text = 'Not Found';
+                break;
+            case 405:
+                $text = 'Method Not Allowed';
+                break;
+            case 406:
+                $text = 'Not Acceptable';
+                break;
+            case 407:
+                $text = 'Proxy Authentication Required';
+                break;
+            case 408:
+                $text = 'Request Time-out';
+                break;
+            case 409:
+                $text = 'Conflict';
+                break;
+            case 410:
+                $text = 'Gone';
+                break;
+            case 411:
+                $text = 'Length Required';
+                break;
+            case 412:
+                $text = 'Precondition Failed';
+                break;
+            case 413:
+                $text = 'Request Entity Too Large';
+                break;
+            case 414:
+                $text = 'Request-URI Too Large';
+                break;
+            case 415:
+                $text = 'Unsupported Media Type';
+                break;
+            case 500:
+                $text = 'Internal Server Error';
+                break;
+            case 501:
+                $text = 'Not Implemented';
+                break;
+            case 502:
+                $text = 'Bad Gateway';
+                break;
+            case 503:
+                $text = 'Service Unavailable';
+                break;
+            case 504:
+                $text = 'Gateway Time-out';
+                break;
+            case 505:
+                $text = 'HTTP Version not supported';
+                break;
+            default:
+                break;
+        }
+        return $text;
+    }
+
+    static function setHeader($code, $message = null)
+    {
+        if ($message == null)
+            $message = self::getMessage($code);
+        $protocol = (isset($_SERVER['SERVER_PROTOCOL']) ? $_SERVER['SERVER_PROTOCOL'] : 'HTTP/1.0');
+        
+        header('HTTP/1.0 ' . $code . ' ' . $text);
+        
+        $GLOBALS['http_response_code'] = $code;
+    }
+}
+?>
