+++ www/client/plugins/filtreapi/ProductHelper.class.php
