+++ www/template/api/card.html
@@ -1,0 +1,1 @@
+#FILTRE_api(card||#ENV{method}||#ENV{product}||#ENV{quantity})
